Camada Controls: Adicionado a classe RegisterCategoryMethods, para que a mesma receba os métodos de seu respectivo form.
